requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
SQLAlchemy==2.0.46
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
APScheduler==3.11.2
python-dotenv==1.2.1

